Add Ctrl+Delete/Backspace to delete words (#1414)
* Add Ctrl+Delete/Backspace to delete words

Equivalent to using Ctrl+Right/Left arrow, but deleting
instead of moving the cursor.

(Note on macOS, Delete is used for deleting backwards,
which is Backspace on other platforms, and fn+Delete
is used for deleting forwards, which is Delete on other
platforms)

Also fixed a few shortcuts that were not set up to work
in the peer text widget: Undo/Redo side-effects (e.g.
when the illo fixup dialog is visible, it should clear and
wait for a re-run after undo/redo); smart cut/copy/paste,
i.e. if column selection, cut/copy/paste the column selection;
smart delete, i.e. if column selection, delete the column
selection. All these worked in the main window, but not
in the peer, since the bindings were not made to the peer
window.

Fixes #1405

* Use correct Mac key for forward/backward word

* Add Option+delete as well as Ctrl+delete

Since Option can be used in some Mac apps when
deleting a word, allow it in GG2.