Fix click-drag bug with NumLock (#1411)
* Fix click-drag bug with NumLock

NumLock was being interpreted as "a modifier key", and
thus causing "copy" behavior.

Fixes #1410

* Minor logic improvement

If copying, allow drag to copy text to location immediately
before/after the selection, but if not copying (or if just
clicking) then releasing mouse in first or last character of
selection now deselects, just like releasing in the middle
of the selection.

* Debug print for Mac test

* Support Cmd-drag for copy text