Sequential Counter (#1417)
* Add `Replace [::] With Incremental Counter`

It's a clone of the GG1 feature. Starting from the top of the
file, it replaces the first `[::]` with `1`, then next with `2`, etc.

Fixes #1370

* Add `cinc` to `\C...\E` functions

Increments & returns the value referenced by the key.

E.g.
`\Ccinc("X")\E` will increment the value stored in X, and
will be replaced by that value. If "X" doesn't exist, it is
assumed to be zero, so the first call will return "1".

Using this as a replacement expression is equivalent to
the `[::]` incremental feature, but more flexible, since it
can easily be made to start at any value, using
`\Ccset("X", 27)\E` as a replacement before all the other
replacements.