Make initial dir persistent for file load/save
File loading, saving and directory choosing dialogs
now remember the last-used directory and open at
that place initially.

Notes:
1. This preference is only set when those dialogs
are used. It is not set if you do `-r1` or similar.
2. There are a couple exceptions where the code
"knows best". If you are using Save As to save a
copy of your file, it opens in the directory of your
current file, not the saved pref directory. Also if
you try to open an image it will try to open in your
pngs directory.
3. The default for a new user is your home directory.
4. If the saved directory doesn't exist, it tries the
parent directory (maybe your projects dir, and you
have deleted the project you just finished with).
5. If the parent directory doesn't exist, it tries the
grandparent directory.
6. If no grandparent, it uses the default (see below)
7. The default in the above case, or for a first-time
user of the feature (or `--nohome`) is the user's
home dir.