Optionally add `h3` markup during autogen
If checkbox is on, detect two blank lines followed by
a single line of text then a blank line, and mark it up
as an `h3` heading.

Multi-line text following 2 blank lines is not treated
as h3, but will just receive `p` markup.

Also, if the single line is an illo, footnote, etc., or
block markup, it will not be converted to `h3`

Fixes #1367