Improve WF diacritics check (#1434)
In `master` an accented word was only suspect if
it also appeared in non-accented form, e.g.
"révolution" and "revolution". The case of
"révolution" and "rèvolution was not flagged.

Now any variation on accenting is flagged.

Also, if all 3 versions:
"revolution", "révolution" and "rèvolution"
occurred, the non-accented form was inserted
twice into the list. Now fixed.

Fixes #1360