Fix SR bugs (#1446)
1. Make Search menu, Find Next/Previous respect the
"Reverse" flag in the S/R dialog. So, Find Next will find
the next in the reverse direction if "Reverse" is set, for
example.
2. Fix exception when Find All is Re-Run after closing
the S/R dialog.
3. Also, set cursor back to non-busy when an
exception happens, because it confuses users into
thinking they need to quit the program, when in
fact they could just carry on.