Display release notes
1. Add "Release Notes" button to Help menu - shows notes
from all releases
2. When program starts, show release notes for current
version if user has not seen them (or any later ones) before.
3. Error if CHANGELOG in `src/guiguts` does not begin with
`# Changelog`
4. Error if first section in CHANGELOG is not `##Version a.b.c`
where `a.b.c` is the current guiguts version.
5. Note that `--nohome` will normally show release notes,
but see 6 below.
6. Don't show release notes if `--debug`/`-d` argument is
given (to help when testing)

Note the above error checks are assertions, so will only
happen to devs or running the test suite. Once packaged
and released to users, the checks won't happen, but the
file will definitely be valid by then.

Also note that this relies on the developer copying the
changelog into the src/guiguts directory when they change
the version number in `pyproject.toml`. If they forget, they
won't be able to merge the version number change because
the test suite will fail, so should be safe.