Review fixes
1. Fixed wording & markup of various tips
2. Changed on/off toggle for tips to Daily/Weekly/Never
radio buttons, which determine whether the tips are shown
when GG starts. These radio buttons appear at the
bottom of the tip dialog, and in the Settings->Advanced tab.
3. When tip is shown (on startup or manually via Help menu)
the current date is stored in the prefs file. A tip will not be
shown on startup unless the current date is 1 day, 7 days,
or 999999 days after the stored date. Date behavior is
probably most easily tested by editing the
`did_you_know_last_shown` value in the Prefs file - must
be a valid date.
4. Make Did You Know, Release Notes, and Help About
Guiguts dialogs all use the label/button font. This is either
the system font or a font family/size chosen by the user
in the Settings->Font tab. The Release Notes and Help
About dialogs both also use 2 slightly larger variants of
the label/button font for headings.