Default to using advanced regex behavior
No real disadvantages, and it saves having to prefix the
regex with `(?V1)` to use any VERSION1 features.

Testing notes:
The regex `[a-z--c-t]` will match all of a-z letters in `master`
whereas with this commit, it will match `abuvwxyz`, i.e. `a-z`
minus `c-t`.
If you want the advanced behavior in `master`, you have to
use `(?V1)[a-z--c-t]`