Cope with `<div>` round Auto-illos
When GG converts `[Illustration]` to HTML, it adds `<p>`
markup around it. Auto-illus removes this when it converts
the illustration to an HTML `<figure>`.

New behavior: If the user customizes that `<p>` markup
with a class, or uses a `<div>` instead, leave their markup
alone and just put the `<figure>` inside it.

Fixes #1512