Cope with `<div>` round Auto-illos (#1514)
* Cope with `<div>` round Auto-illos

When GG converts `[Illustration]` to HTML, it adds `<p>`
markup around it. Auto-illus removes this when it converts
the illustration to an HTML `<figure>`.

New behavior: If the user customizes that `<p>` markup
with a class, or uses a `<div>` instead, leave their markup
alone and just put the `<figure>` inside it.

Fixes #1512

* Cope if user moves spotlit text

If user edited class name, for example, they might move the
spotlit text, so re-grab it just before converting to HTML.

* If currently at illo, go to Next

Instead of reselecting the same one.