Allow "Per LZ" re-indexing of footnotes
User sets up Landing Zones, normally either at end of book,
or end of chapter, but maybe end of section, or other scheme.
Checkbox next to Reindex button causes renumbering to
start from 1 for each LZ.
Note that the check for duplicate footnote numbers is now
only done when "Per LZ" is turned off.

The HTML generation hasn't changed - it already used 2
numbers, the first is the label of the footnote, and the
second is the number of the footnote counting from the
start of the book. Therefore, FNs will get IDs something
like: 1_1, 2_2, 3_3, 1_4, 2_5, 1_6, etc. They will thus be unique
which is the only requirement.

Fixes #526
Fixes #909 (well, sort-of - it makes it less bizarre)