Cope with multiple anchors for a footnote
Instead of saying it's found a duplicate or unmatched anchor
number, it now links both anchors with the same footnote.
It also doesn't report the anchors being out of order, since
that is now permitted (not going to make the logic more
complicated than it already is)

Testing notes:
Hopefully I've trapped all the places that needed to be able
to cope with more than one anchor.

Please check that the anchors are reported correctly in the
FN fixup dialog; reindex works; reindex per LZ works; move
to end LZ or chapter LZ work; move to paragraph works.