Add `--resetgeometry` command line option (#1527)
No PPers have reported a problem with multiple monitors,
but if the main window or dialogs are dragged onto a second
monitor, then the monitor is removed before the next run of
GG, the windows may be placed off-screen.

Use `guiguts --resetgeometry` to reset the size and position
of the main window and all dialogs to the default values,
which will bring them back onto the main screen.

I think this is sufficient for now for a situation that hasn't
arisen during PPing.

Fixes #206