Allow split window viewing of footnotes (#1534)
* Allow split window viewing of footnotes

1. Add preference to Settings dialog, Advanced tab. (Note
some "advanced" settings have been moved to the
"appearance" tab to make space
2. If pref is off, or the window is not split, behavior should be
unchanged.
3. If pref is on and window is split, when a footnote or anchor
message is selected in the FN dialog, the anchor is shown
in the top window, and the FN in the bottom.
4. Both the FN and the anchor are highlighted - note that
highlights will show in both windows, so if the anchor is close
to the FN, you may see them both in both windows.

Fixes #859

* Make Hide All & Shift-right click hide by type

Either hide all footnotes or hide all anchors, depending on
selected message.