Improve support for Google Ngram Viewer (#1560)
1. Add `Google Ngram parameters` field to Settings,
Advanced tab, where user can specify corpus, start year,
end year, etc. Default to use English 2019 dataset.
2. Remove Ngram Viewer from Custom Menu if the label
string and command are still set to the default values. If
user has customized the entry, do not remove it.
3. Add `Google Books Ngram Viewer` to Tools Menu.
4. If user has text selected, pass that to the Ngram Viewer.
5. If user has selected a single, hyphenated word, pass that
word and its unhyphenated form to the Ngram Viewer
6. In Word Frequency, make Shift-Cmd/Ctrl-Left-Click show
word in Ngram Viewer.
7. If WF is in "Hyphens" mode, and selected word is marked
as a suspect, pass the word and all the other suspects in
the viewer. This could be "flash-light" & "flashlight", or
"flash-light" & "flash light", or all 3 variants, depending on
what is in the text.
8. If in WF Hyphens mode, but word is not marked as a
suspect, it must contain a hyphen, so pass the word and
its unhyphenated form to the Ngram viewer.

Fixes #1559