Report CP Filter effects in more detail (#1562)
* Report CP Filter effects in more detail

1. In CP Filter dialog, display list of changes made
2. Verbose flag determines whether a single line is
summarizing the number of changes of each type,
or whether each line that is changed is reported, with the
line number, so user can click and visit it.

* Fix message reselection bug

When Fixing an error in an "ALL_MESSAGES" matching
dialog, it forgot which message was selected, and after
fixing selected the first instead.

Introduced by #1554