Strengthen warning about astral plane characters
In Unicode Search dialog, and in the description at the bottom
of the Unicode block dialog when hovering over a character,
it used ⚠ to indicate it was a "new" character, which should
be used with caution. This commit adds ☢ to indicate if
it's an astral plane character and should not be used.
Tooltip of Unicode Search dialog updated accordinly.

Fixes #1596