Change Find All matching to highlighted text
Previously Hide All only hid messages in the Find All dialog
if they were identical to the selected one. This commit makes
it hide messages where the highlighted (matched) string is
the same.

This means if a user does Find All with a regex, and the
regex matches several different strings, they can hide all the
messages that relate to one of the strings.