Fix reindex bug if 2 anchors per footnote (#1598)
Bug occurred because it changed the number in the footnote
for every anchor. So if there were two anchors, it changed
the number in the footnote twice. However, the change was
based on character positions, so if the number was "26", it
replaced those two characters with "XXVI.", for example, then
when it did the second replacement, it replaced two of
those characters with "XXVI." leading to something like
"XXVI.VI."

Fixes #1595