Add "swap" buttons to CP Olde Englifh dialog (#1607)
* Add "swap" buttons to CP Olde Englifh dialog

Allow swapping of "f", "s" and "*"

Will not be able to cope with all situations where two
characters are in the word, e.g. swapping "f" for "s" in
"foreft" gives "sorest", but can't get "forest".

Fixes #1605

* Make Englifh Fix All only fix same "type"

Type is either "*" (in red)  or "s" (in green) which describes
what the "f" will be replaced with.

User can now do Fix All on a green message to fix all the
unambiguous ones, then work more manually. Can also
sort by Alpha/Type to separate the messages into reds
and greens.