Preserve newlines in source of  HTML Autotable
For a Multiline table, chunks of text were concatenated with
a space between to construct the `<td>` element, thus
losing the newlines from the original. Instead concatenate
them with newlines between to preserve text structure.
This has no effect on the book as viewed in the browser,
only the HTML source.